+++ libs/providers/langchain-google-gauth/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @langchain/google-gauth
 
+## 2.1.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @langchain/google-common@2.1.7
+
 ## 2.1.6
 
 ### Patch Changes