+++ libs/providers/langchain-google-vertexai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @langchain/google-vertexai
 
+## 2.1.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @langchain/google-gauth@2.1.7
+
 ## 2.1.6
 
 ### Patch Changes